made the course route restful

+++ config/cloudinary.js
@@ -1,28 +1,32 @@
-import { v2 as cloudinary} from "cloudinary";
+import { v2 as cloudinary } from "cloudinary";
 import fs from "fs";
 
-const uploadOnCludinary = async (filePath) => {
+const uploadOnCloudinary = async (filePath) => {
   cloudinary.config({
     cloud_name: process.env.CLOUDINARY_NAME,
     api_key: process.env.CLOUDINARY_API_KEY,
     api_secret: process.env.CLOUDINARY_API_SECRET,
   });
+
   try {
-    
-    if (!filePath) {
-      return null;
-    }
+    if (!filePath) return null;
+
     const uploadResult = await cloudinary.uploader.upload(filePath, {
       resource_type: "auto",
     });
+
     if (fs.existsSync(filePath)) {
-      fs.unlinkSync(filePath); 
+      fs.unlinkSync(filePath);
     }
-    return uploadResult.secure_url;
+    
+    return uploadResult.secure_url; 
   } catch (error) {
-    fs.unlinkSync(filePath);
-    console.log(error);
+    if (filePath && fs.existsSync(filePath)) {
+        fs.unlinkSync(filePath);
+    }
+    console.error("Cloudinary Error:", error);
+    return null;
   }
 };
 
-export default uploadOnCludinary
+export default uploadOnCloudinary;
